Win build fix: doesn't have ssize_t WTF?

+++ lib/gui_rpc_client.cpp
@@ -309,7 +309,7 @@ int RPC_CLIENT::send_request(const char* p) {
     buf = "<boinc_gui_rpc_request>\n";
     buf += p;
     buf += "</boinc_gui_rpc_request>\n\003";
-    ssize_t n = send(sock, buf.c_str(), (int)buf.size(), 0);
+    int n = (int)send(sock, buf.c_str(), (int)buf.size(), 0);
     if (n < 0) {
         //printf("send: %d\n", n);
         //perror("send");
